New query demonstrating controlling OR precedence.
The prior commit used a query that coincidentally worked out because of
the SQL standard defining AND at a higher precedence than OR.
ActiveRecord did not make any effort to group the expressions,
presumably because it knew it wasn't necessary.

The query in this commit requires the OR to have higher precedence,
thus requires parentheses around that portion of the WHERE clause.
Adding in the `or` portion via a `merge` method ensures that the
grouping is maintained properly, and the output SQL is correct.